pref: merge two cron events
