+++ README.md
@@ -45,6 +45,8 @@ Astro 提供了很多部署方案，自行 [查看设置方法](https://docs.ast
 | `OPENAI_API_BASE_URL`              | 本地开发时可以填写 OpenAI 的代理服务器，但是 Vercel 不需要。 | api.openai.com                                               |
 | `PASSWORD`                         | 网站密码                                                     | 无                                                           |
 | `MAX_INPUT_TOKENS`                 | 输入的 token 最大值，如果开启 `连续对话`，将计算之前的所有对话内容。OpenAI 限制 token 最大值为 4096，但这是输入和输出之和，所以可以将这个值设置为 3072， 留 1024 作为输出。如果不想被滥用，可以将这个值设置的再小一点。 | 3072                                                         |
+| `SENDKEY`                          | 使用 [Server酱](https://sct.ftqq.com/sendkey) 推送帐号余额以及被 ban 状态到微信，如果需要自行获取。推送事件为 早上 8 点和晚上 8 点，在 vercel.json 文件中修改。如果 key 太多，超过 20 个，有超时的可能。 | 无                                                           |
+| `SENDCHANNEL`                      | [Server酱](https://sct.ftqq.com/sendkey) 的推送通道，默认微信服务号。 | 9                                                            |
 
 有两种设置方式
 1. 将 `.env.example` 文件修改为 `.env`，在 `.env` 中设置。
